requirements.txt
streamlit
matplotlib
pandas
seaborn
wordcloud
pillow
numpy
snowflake-snowpark-python